Create `url` module shared between crates
